requirements.txt
pandas==1.5.3
google-api-python-client==2.70.0
google-auth==2.16.2
openpyxl==3.1.1
